refactor: FilterModalが自分でvillageからparticipantsを導出 (step-7)
- FilterParticipant型と変換ロジックをFilterModal内に移動
- participants/myselfId/notificationKeyword propsを削除しhook/contextで取得
- participants.tsからFilterParticipant関連コードを削除
